view model adjustments
