+++ UpcomingMovies/UpcomingMovies/MovieListViewModel.swift
@@ -8,7 +8,6 @@
 
 import Foundation
 import SwiftUI
-import Combine
 
 public class MoviesViewModel{
     private let movieModels : Movies
@@ -24,62 +23,8 @@ public class MoviesViewModel{
         }
         return moviesVM
     }
-}
-
-public class MovieViewModel: Identifiable{
-    private let movie : Movie
     
-    init(movie: Movie){
-        self.movie = movie
+    var buttonText : String{
+        return self.movies.isEmpty ? "Try Again" : "Load More"
     }
-    
-    public var id : Int? {
-        return movie.id
-    }
-
-    var title : String {
-        return (movie.title ?? "no title")
-    }
-    
-    var adultText : String {
-        guard let adult = movie.adult else{
-            return "No age information"
-        }
-        if adult{
-            return "Adult Movie 🔞"
-        }else{
-            return ""
-        }
-    }
-    
-    var posterPath : URL {
-        return movie.getAbsolutePosterURL()
-    }
-    
-    var backdropPath : URL {
-        return movie.getAbsoluteBackdropURL()
-    }
-    
-    var releaseDate : String {
-        return movie.releaseDate ?? "unknown release date"
-    }
-    
-    func getGenreTextListForCodes(genres: Genres) -> String{
-        var resultString = ""
-        var isFirst = true
-        for code in self.movie.genreIDS ?? []{
-            for genre in genres{
-                if code == genre.id{
-                    if isFirst{
-                        resultString.append(genre.name ?? "")
-                        isFirst = false
-                    }else{
-                        resultString.append(", \(genre.name ?? "")")
-                    }
-                }
-            }
-        }
-        return resultString
-    }
-    
 }
